Improve vault entry creation
Appropriate inputs are now autofocused. When creating a new vault entry, an initial password is generated, and in the browser extension, the title and URL of the entry are filled in based on the URL of the current tab. Reduced the minimum master password length to 12. Vault page state now persists across browser extension launches.