Fix bugs
Fixed a bug where autofill would initially be unavailable after logging in to the extension. Fixed a bug with master password input. Fixed a regression that caused logins to not redirect to the app. Fixed the Mac and Windows window controls being swapped. Improved error handling. Removed CORS in favor of it being handled by NGINX.