Make the theme setting per-device
The theme setting is now per-device rather than synchronized. Fixed a regression where the inputs on the login screen no longer autofocused. Fixed a bug where the biometric login option would appear on the web app settings page when viewed on a mobile device. Refactored the "Options" variables file to be called "Constants".