Redo the browser extension
Redid the browser extension to make the background script lighter and ease the transition to manifest v3. Moved default account selection to the main application. Fixed a bug where logging out and then back in with a different account would fail. Added a temporary workaround for a Sass regression which affected the appearance of the window controls and switches. Updated the icons.